Задача: Считаем прогулы. Создание класса StudentService для инкапсуляции логики управления списком студентов. Удаление лишней HashMap для упрощения логики.
