+++ src/main/java/school/faang/countabsenteeism/Student.java
@@ -1,10 +1,12 @@
-package school.faang.count_absenteeism;
+package school.faang.countabsenteeism;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
+@Setter
 @AllArgsConstructor
 @EqualsAndHashCode
 public class Student {
