+++ src/main/java/school/faang/countabsenteeism/StudentProfile.java
@@ -1,11 +1,13 @@
-package school.faang.count_absenteeism;
+package school.faang.countabsenteeism;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import lombok.Setter;
 import lombok.ToString;
 
 @Getter
+@Setter
 @AllArgsConstructor
 @ToString
 @EqualsAndHashCode
