+++ src/main/java/school/faang/countabsenteeism/StudentService.java
@@ -1,34 +1,13 @@
-package school.faang.count_absenteeism;
+package school.faang.countabsenteeism;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class Main {
-    public static final HashMap<Student, Integer> INDEX_STUDENT = new HashMap<>();
-    public static final List<Student> STUDENTS = new ArrayList<>();
+import static school.faang.countabsenteeism.Main.STUDENTS;
 
-    public static void main(String[] args) {
-        add(new Student("Dmitry", new StudentProfile("Humanities", 1)));
-        add(new Student("Yaroslav", new StudentProfile("Mathematics", 3)));
-        add(new Student("Sergei", new StudentProfile("Humanities", 1)));
-        add(new Student("Alexey", new StudentProfile("Humanities", 3)));
-        add(new Student("Alena", new StudentProfile("Chemistry", 2)));
-        add(new Student("Anastasia", new StudentProfile("Chemistry", 2)));
-        add(new Student("Anatoly", new StudentProfile("Chemistry", 2)));
-        add(new Student("Ruslan", new StudentProfile("Chemistry", 2)));
-
-        remove("Yaroslav", "Mathematics", 3);
-        remove("Dmitry", "Humanities", 1);
-
-
-        List<Student> testGetStudents = getStudentsByStudentProfile("Chemistry", 2);
-        printForTest(testGetStudents);
-
-        printAll(STUDENTS);
-    }
-
+public class StudentService {
     public static HashMap<StudentProfile, ArrayList<Student>> groupingStudents(List<Student> students) {
         HashMap<StudentProfile, ArrayList<Student>> groupedStudents = new HashMap<>();
         for (Student student : students) {
@@ -40,15 +19,13 @@ public class Main {
 
     public static void add(Student student) {
         STUDENTS.add(student);
-        INDEX_STUDENT.put(student, STUDENTS.indexOf(student));
     }
 
     public static void remove(String name, String faculty, int year) {
         Student studentToRemove = new Student(name, new StudentProfile(faculty, year));
-        int indexStudent = INDEX_STUDENT.get(studentToRemove);
+        int indexStudent = STUDENTS.indexOf(studentToRemove);
         if (studentToRemove != null) {
             STUDENTS.remove(indexStudent);
-            INDEX_STUDENT.remove(studentToRemove);
         }
     }
 
@@ -56,7 +33,7 @@ public class Main {
         ArrayList<Student> studentsByStudentProfile = new ArrayList<>();
         for (Student student : STUDENTS) {
             if (student.getStudentProfile().getFaculty().equals(faculty)
-            && student.getStudentProfile().getYear() == year) {
+                    && student.getStudentProfile().getYear() == year) {
                 studentsByStudentProfile.add(student);
             }
         }
